Project Development Phase/Sprint 2/requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
astunparse==1.6.3
autopep8==2.0.0
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
gevent==22.10.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.50.0
h5py==3.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mysqlclient==2.1.1
nltk==3.7
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.2
